Stop erroring if every transaction change is zero
In some situations, there can be a zero-value transaction in the
imported CSV files. These are usually for card checks and similar
usecases. Prior to this change, the ledger tool would return the `more
than one account empty` error in that case.

This change updates the logic to return early if the balances sum to
zero. If there is any remaining balance, then the old behaviour is
preserved.